backend/requirements.txt
flask
flask-cors
google-generativeai
Pillow
Werkzeug
python-dotenv
gunicorn
pymongo
